U change code to rubycode style :-)
